+++ decide.c
@@ -59,6 +59,10 @@ random_decision (struct ring *ring)
 
   LOG ("random decision %s", LOGVAR (idx));
 
+  if (ring->context == SEARCH_CONTEXT)
+    ring->statistics.decisions.random++;
+
+
   return idx;
 }
 
@@ -88,6 +92,9 @@ best_decision_on_heap (struct ring *ring)
   LOG ("best decision %s on heap with score %g",
        LOGVAR (idx), nodes[idx].score);
 
+  if (ring->context == SEARCH_CONTEXT)
+    ring->statistics.decisions.heap++;
+
   return idx;
 }
 
@@ -115,6 +122,10 @@ best_decision_on_queue (struct ring *ring)
 
   LOG ("best decision %s on queue with stamp %" PRIu64,
        LOGVAR (idx), search->stamp);
+
+  if (ring->context == SEARCH_CONTEXT)
+    ring->statistics.decisions.queue++;
+
   return idx;
 }
 
@@ -135,8 +146,17 @@ decide (struct ring *ring)
 
   signed char phase = decide_phase (ring, idx);
   unsigned lit = LIT (idx);
+
   if (phase < 0)
     lit = NOT (lit);
+
+  if (ring->context == SEARCH_CONTEXT)
+    {
+      if (phase < 0)
+	ring->statistics.decisions.negative++;
+      else
+	ring->statistics.decisions.positive++;
+    }
 
   ring->level++;
   assign_decision (ring, lit);